requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
Faker==4.1.1
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pg==0.1
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.3
PyOpenGL==3.1.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
validate-docbr==1.7.0
whitenoise==5.2.0
wrapt==1.12.1